requirements.txt
absl-py==2.4.0
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==3.25.2
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
deepface==0.0.98
executing==2.2.1
facenet-pytorch==2.6.0
filelock==3.24.3
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fsspec==2026.2.0
gast==0.7.0
gdown==5.2.1
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.1.0
h5py==3.15.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==2.14.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==1.0.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9.1
rubicon-objc==0.5.3
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==1.14.2
